# src/main/resources/luckperms-openapi.yml
openapi: 3.0.0
info:
  title: LuckPerms REST API
  description: The LuckPerms REST API allows developers to query LuckPerms data using HTTP calls.
  version: '1.0'
  contact:
    name: LuckPerms Team
    url: 'https://luckperms.net'
  license:
    name: MIT
    url: 'https://github.com/LuckPerms/LuckPerms/blob/master/LICENSE.txt'
servers:
  - url: ../
...
paths:
  /user:
    get:
      summary: Get all existing users
      operationId: get-users
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                uniqueItems: true
                items:
                  $ref: '#/components/schemas/UniqueId'
              examples:
                example-1:
                  value:
                    - c1d60c50-70b5-4722-8057-87767557e50d
      description: Returns an array of all known users.
      tags:
        - Users
    post:
      summary: Create a new user
      operationId: create-user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerSaveResult'
              examples:
                example-1:
                  value:
                    outcomes:
                      - username_updated
                      - other_unique_ids_present_for_username
                    previousUsername: Luck
                    otherUniqueIds:
                      - c1d60c50-70b5-4722-8057-87767557e50d
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerSaveResult'
              examples:
                example-1:
                  value:
                    outcomes:
                      - clean_insert
        '400':
          description: Missing required information
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
            examples:
              example-1:
                value:
                  uniqueId: c1d60c50-70b5-4722-8057-87767557e50d
                  username: Luck
        description: ''
      description: Create a new user.
      tags:
        - Users
  /user/lookup:
    get:
      summary: Search for a user with the given username or unique id
      operationId: get-user-lookup
      tags:
        - Users
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  uniqueId:
                    $ref: '#/components/schemas/UniqueId'
                  username:
                    type: string
              examples:
                example-1:
                  value:
                    uniqueId: c1d60c50-70b5-4722-8057-87767557e50d
                    username: Luck
        '400':
          description: Missing required information
        '404':
          description: User doesn't exist
      description: |-
        Lookup the unique id or username of a user with the given username or unique id.

        You must specify one of the query parameters in the request.
      parameters:
        - schema:
            type: string
            minLength: 1
            example: Luck
          in: query
          name: username
          description: The username to search for
        - schema:
            type: string
            format: uuid
            pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
            example: c1d60c50-70b5-4722-8057-87767557e50d
          in: query
          name: uniqueId
          description: The unique id to search for
    parameters: []
  /user/search:
    get:
      summary: Search for users with given nodes
      tags:
        - Users
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSearchResult'
              examples:
                example-1:
                  value:
                    - uniqueId: c1d60c50-70b5-4722-8057-87767557e50d
                      results:
                        - key: minecraft.command.ban
                          type: permission
                          value: true
                          context:
                            - key: server
                              value: survival
                          expiry: 1658428395
      operationId: get-user-search
      description: |-
        Search for users with given nodes.

        You must specify one of the query parameters in the request.
      parameters:
        - schema:
            type: string
            minLength: 1
          in: query
          name: key
          description: Search for nodes with a key equal to
        - schema:
            type: string
            minLength: 1
          in: query
          name: keyStartsWith
          description: Search for nodes with a key starting with
        - schema:
            type: string
            minLength: 1
          in: query
          name: metaKey
          description: Search for meta nodes with a meta key equal to
        - schema:
            type: string
            enum:
              - regex_permission
              - inheritance
              - prefix
              - suffix
              - meta
              - weight
              - display_name
          in: query
          name: type
          description: Search for nodes with a type equal to
        - schema:
            type: string
            minLength: 1
          in: query
          name: group
          description: Search for inheritance nodes with the given group name
  '/user/{uniqueId}':
    parameters:
      - $ref: '#/components/parameters/uniqueId'
    get:
      summary: Get a users data
      tags:
        - Users
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              examples:
                example-1:
                  value:
                    uniqueId: c1d60c50-70b5-4722-8057-87767557e50d
                    username: Luck
                    nodes:
                      - key: group.default
                        type: inheritance
                        value: true
                        context: []
                      - key: minecraft.command.ban
                        type: permission
                        value: true
                        context:
                          - key: server
                            value: survival
                        expiry: 1658428395
                    metadata:
                      meta:
                        nametag-color: red
                        fly-speed: '5'
                      prefix: '[Admin] '
                      suffix: ' [the great]'
                      primaryGroup: default
        '404':
          description: User doesn't exist
      operationId: get-user
      parameters:
        - $ref: '#/components/parameters/uniqueId'
      description: Get a user by unique id (UUID).
    patch:
      summary: Update a users data
      operationId: patch-user
      responses:
        '200':
          description: OK
        '400':
          description: Missing required information
        '404':
          description: User doesn't exist
      description: Update a user's data
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
                - properties:
                    username:
                      type: string
                      minLength: 1
                      example: Luck
              type: object
            examples:
              example-1:
                value:
                  username: Luck
        description: ''
      tags:
        - Users
    delete:
      summary: Delete a user
      operationId: delete-user
      responses:
        '200':
          description: OK
        '404':
          description: User doesn't exist
      description: Delete a user
      parameters:
        - schema:
            type: boolean
          in: query
          name: playerDataOnly
          description: if only player data should be deleted
          required: false
      tags:
        - Users
...
  /group:
    get:
      summary: Get all existing groups
      operationId: get-groups
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                uniqueItems: true
                items:
                  $ref: '#/components/schemas/GroupName'
              examples:
                example-1:
                  value:
                    - admin
      description: Get all known groups
      tags:
        - Groups
    post:
      summary: Create a new group
      operationId: create-group
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
              examples:
                example-1:
                  value:
                    name: admin
                    nodes: []
                    metadata:
                      meta: {}
        '400':
          description: Missing required information
        '409':
          description: Group already exists
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewGroup'
            examples:
              example-1:
                value:
                  name: admin
        description: ''
      description: Create a new group.
      tags:
        - Groups
  /group/search:
    get:
      summary: Search for groups with given nodes
      tags:
        - Groups
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupSearchResult'
              examples:
                example-1:
                  value:
                    - name: admin
                      results:
                        - key: minecraft.command.ban
                          type: permission
                          value: true
                          context:
                            - key: server
                              value: survival
                          expiry: 1658428395
      operationId: get-group-search
      description: |-
        Search for groups with given nodes.

        You must specify one of the query parameters in the request.
      parameters:
        - schema:
            type: string
            minLength: 1
          in: query
          name: key
          description: Search for nodes with a key equal to
        - schema:
            type: string
            minLength: 1
          in: query
          name: keyStartsWith
          description: Search for nodes with a key starting with
        - schema:
            type: string
            minLength: 1
          in: query
          name: metaKey
          description: Search for meta nodes with a meta key equal to
        - schema:
            type: string
            enum:
              - regex_permission
              - inheritance
              - prefix
              - suffix
              - meta
              - weight
              - display_name
          in: query
          name: type
          description: Search for nodes with a type equal to
        - schema:
            type: string
            minLength: 1
          in: query
          name: group
          description: Search for inheritance nodes with the given group name
  '/group/{groupName}':
    parameters:
      - $ref: '#/components/parameters/groupName'
    get:
      summary: Get a groups data
      operationId: get-group
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
              examples:
                example-1:
                  value:
                    name: admin
                    displayName: Admin
                    weight: 10
                    nodes:
                      - key: minecraft.command.ban
                        type: permission
                        value: true
                        context:
                          - key: server
                            value: survival
                        expiry: 1658428395
                    metadata:
                      meta:
                        nametag-color: red
                        fly-speed: '5'
                      prefix: '[Admin] '
                      suffix: ' [the great]'
                      primaryGroup: admin
        '404':
          description: Group doesn't exist
      description: Get a group
      tags:
        - Groups
    delete:
      summary: Delete a group
      operationId: delete-group
      responses:
        '200':
          description: OK
        '404':
          description: Group doesn't exist
      description: Delete a group
      tags:
        - Groups
...
  /track:
    get:
      summary: Get all existing tracks
      operationId: get-tracks
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                uniqueItems: true
                items:
                  $ref: '#/components/schemas/TrackName'
              examples:
                example-1:
                  value:
                    - staff
      description: Get all known tracks
      tags:
        - Tracks
    post:
      summary: Create a new track
      operationId: create-track
      responses:
        '201':
          description: Track created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
              examples:
                example-1:
                  value:
                    name: staff
                    groups:
                      - moderator
                      - admin
                      - owner
        '400':
          description: Missing required information
        '409':
          description: Track already exists
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTrack'
            examples:
              example-1:
                value:
                  name: staff
        description: ''
      description: Create a new track.
      tags:
        - Tracks
  '/track/{trackName}':
    parameters:
      - $ref: '#/components/parameters/trackName'
    get:
      summary: Get a tracks data
      operationId: get-track
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
              examples:
                example-1:
                  value:
                    name: staff
                    groups:
                      - moderator
                      - admin
                      - owner
        '404':
          description: Track doesn't exist
      description: Get a track
      tags:
        - Tracks
    patch:
      summary: Update a track
      operationId: patch-track
      responses:
        '200':
          description: OK
        '400':
          description: Missing required information
        '404':
          description: Track doesn't exist
      description: Update a track
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
                - properties:
                    groups:
                      type: array
                      items:
                        $ref: '#/components/schemas/GroupName'
              type: object
            examples:
              example-1:
                value:
                  groups:
                    - moderator
                    - admin
                    - owner
        description: ''
      tags:
        - Tracks
    delete:
      summary: Delete a track
      operationId: delete-track
      responses:
        '200':
          description: OK
        '404':
          description: Track doesn't exist
      description: Delete a track
      tags:
        - Tracks
  /action:
    get:
      summary: Query actions
      operationId: get-actions
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/Action'
                  overallSize:
                    type: integer
              examples:
                example-1:
                  value:
                    entries:
                      - timestamp: 1658428395
                        source:
                          uniqueId: c1d60c50-70b5-4722-8057-87767557e50d
                          name: Luck
                        target:
                          uniqueId: 069a79f4-44e9-4726-a5be-fca90e38aaf5
                          name: Notch
                          type: user
                        description: permission set minecraft.command.ban true
                    overallSize: 1
        '400':
          description: Missing required information
      description: |
        Query actions from the action logger.
        
        If pageSize or pageNumber are specified, both must be specified.
        If neither are specified, no pagination will be used and all results will be returned.
      parameters:
        - schema:
            type: integer
            minimum: 1
          in: query
          name: pageSize
          description: The number of actions to return on each page
        - schema:
            type: integer
            minimum: 1
          in: query
          name: pageNumber
          description: The page to return
        - schema:
            type: string
            minLength: 1
          in: query
          name: source
          description: Filter by source user unique id
        - schema:
            type: string
            minLength: 1
          in: query
          name: user
          description: Filter by target user unique id
        - schema:
            type: string
            minLength: 1
          in: query
          name: group
          description: Filter by target group name
        - schema:
            type: string
            minLength: 1
          in: query
          name: track
          description: Filter by target track name
        - schema:
            type: string
            minLength: 1
          in: query
          name: search
          description: Filter by search value in source name, target name or description.
      tags:
        - Actions
    post:
      summary: Submit a new action
      operationId: submit-action
      responses:
        '202':
          description: Accepted
        '400':
          description: Missing required information
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Action'
      description: Submit a new action to the action logger.
      tags:
        - Actions
  /messaging/update:
    post:
      summary: Push an update via the messaging service
      operationId: messaging-push-update
      responses:
        '202':
          description: Accepted
        '501':
          description: Messaging service not supported
      description: Push an update via the messaging service.
      tags:
        - Messaging
...
  /messaging/custom:
    post:
      summary: Send a custom message via the messaging service
      operationId: messaging-custom-message
      responses:
        '202':
          description: Accepted
        '501':
          description: Messaging service not supported
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomMessage'
      description: Send a custom message via the messaging service.
      tags:
        - Messaging
...
components:
  schemas:
...
    User:
      title: User
      type: object
      properties:
        uniqueId:
          $ref: '#/components/schemas/UniqueId'
        username:
          type: string
          minLength: 1
          example: Luck
          description: |-
            The username of the user.

            Usernames are normalised to lower case by the platform.
        parentGroups:
          type: array
          items:
            $ref: '#/components/schemas/GroupName'
        nodes:
          $ref: '#/components/schemas/NodeMap'
        metadata:
          $ref: '#/components/schemas/Metadata'
      required:
        - uniqueId
        - nodes
        - metadata
    NewUser:
      title: NewUser
      type: object
      properties:
        uniqueId:
          $ref: '#/components/schemas/UniqueId'
        username:
          type: string
          minLength: 1
          example: Luck
      required:
        - uniqueId
        - username
...
    UniqueId:
      type: string
      title: Unique Id
      description: A player unique id (UUID)
      example: c1d60c50-70b5-4722-8057-87767557e50d
      format: uuid
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
...
    GroupName:
      type: string
      title: Name
      description: A group name
      example: admin
      pattern: '^[a-z0-9]+$'
      minLength: 1
...
    UserSearchResult:
      title: UserSearchResult
      type: object
      properties:
        uniqueId:
          $ref: '#/components/schemas/UniqueId'
        results:
          $ref: '#/components/schemas/NodeMap'
      description: ''
...
    PlayerSaveResult:
      title: PlayerSaveResult
      type: object
      description: The result of an operation to save data about a player
      properties:
        outcomes:
          type: array
          description: the outcomes of the operation
          uniqueItems: true
          items:
            $ref: '#/components/schemas/PlayerSaveResultOutcome'
        previousUsername:
          type: string
          description: the previous username involved in the result (only applies for the username_updated outcome)
          minLength: 1
          example: Luck
        otherUniqueIds:
          type: array
          description: the other unique ids involved in the result (only applies for the other_unique_ids_present_for_username outcome)
          items:
            $ref: '#/components/schemas/UniqueId'
      required:
        - outcomes
...
  securitySchemes:
    apikey:
      type: http
      scheme: bearer
      description: API Key used to authenticate
...
  parameters:
    uniqueId:
      name: uniqueId
      in: path
      schema:
        $ref: '#/components/schemas/UserIdentifier'
      required: true
      description: |-
        A player identifier.

        Either a unique id (UUID), or a username. If the value isn't a UUID, it's looked up as a username, and a
        `404` is returned if no user has it.
...
security:
  - apikey: []